myapp/requirements.txt
gradio==5.23.2
transformers==4.38.2
torch==2.2.1
Pillow==10.3.0
